+++ evm/src/main/java/org/hyperledger/besu/evm/fluent/SimpleAccount.java
@@ -114,8 +114,8 @@ public class SimpleAccount implements MutableAccount {
   }
 
   @Override
-  public Optional<Integer> getCodeSize() {
-    return Optional.of(this.code == null ? 0 : this.code.size());
+  public Optional<Long> getCodeSize() {
+    return Optional.of((long) (this.code == null ? 0 : this.code.size()));
   }
 
   @Override
